requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
anthropic==0.7.8
docker==6.1.3
python-multipart==0.0.6
aiohttp==3.9.1
asyncio-mqtt==0.13.0
python-dotenv==1.0.0
